[thirdparty] explicit include directive for gumbo-query
There were reports that building Kudu as a part of Impala's toolchain
on Ubuntu 24.04 LTS would fail on compiling gumbo-query because the
necessary include directory directive -I$PREFIX/include was not added
by CMake, while it should have been present.  Getting to the bottom
of the issue might be time-consuming, but adding the include directory
explicitly is easy and shouldn't hurt since we already do so when
building glog at least.

Change-Id: I65b8a6d3b63ff6684e2ed8579b01f25c243ffc66
Reviewed-on: http://gerrit.cloudera.org:8080/23501
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Tested-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Gabriella Lotz <lotzgabriella@gmail.com>