dist_test: Add directory support for KUDU_DATA_FILES
Modified dist_test.py to handle directories in addition to files when
processing KUDU_DATA_FILES. When a data file path resolves to a
directory, append a trailing slash so the isolate tool includes it
recursively.

This change is needed for the Swagger documentation patch, which
requires setting the document root to include the entire www/
directory structure.

Change-Id: Ib1f88760f43c1cacb538f64454c6d3657295742f
Reviewed-on: http://gerrit.cloudera.org:8080/23538
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Tested-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Zoltan Martonka <zmartonka@cloudera.com>