KUDU-1261 array columns aren't yet supported as key columns
This changelist adds a guardrail to explicitly disallow creating Kudu
tables with array key columns.  This artificial restriction is added
to prevent unexpected behavior since a proper encoding for array column
cells isn't implemented yet.  A new unit test to verify the expected
behavior is added as well.

As a bit of context, to properly encode an array cell, it's necessary
to make sure that:
  * the contents of the memory buffer backing the cell is 'normalized',
    i.e. the interpretation of null/non-valid elements of a particular
    type is always the same when feeding data to the encoder
  * run the encoder against the sub-range of the memory buffer that
    contains array elements' data only, and not include metadata, etc.

Change-Id: If37a4dd2d1689aa51ab09e6cb71f01664dc2ee1a
Reviewed-on: http://gerrit.cloudera.org:8080/23552
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Ashwani Raina <araina.git@gmail.com>
Tested-by: Alexey Serbin <alexey@apache.org>