KUDU-1261 introduce ARRAY_1D_COLUMN_TYPE feature for tservers
This changelist introduces a new feature flag ARRAY_1D_COLUMN_TYPE
for Kudu tablet servers to reflect whether a tablet server is capable
of hosting tablets with array columns.  Also, the system catalog's code
has been updated to require the feature flag when sending requests
to create a table with array type columns or altering a table
that contains at least one array type column.

The rationale behind adding the new feature flag is to make it easier
to reason about and troubleshoot situations when a Kudu master of newer
version works with tablet servers of prior versions, and the latter ones
do not support array type columns.

This patch includes test scenarios to provide basic coverage
for the newly introduced functionality.

As a part of this changelist, I also updated the logic of the catalog
manager's RetryingTSRpcTask::RpcCallback() method to not retry sending
same RPC to tablet servers again and again if the tablet server responds
with 'unsupported feature flags'.  This behavior is gated by the newly
introduced flag which is set to 'false' by default:
  --catalog_manager_retry_tserver_task_on_unsupported_feature_flags

Change-Id: I07a8e6b6a90a90562e9ad0be35702f21b49d2da5
Reviewed-on: http://gerrit.cloudera.org:8080/23559
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>