KUDU-1261 use empty validity vector when no nulls in array
This changelist updates conventions on supplying and consuming the
validity information on array elements in the Kudu C++ client API.
With this changelist, an empty validity vector is accepted if all
elements in data array to be written are non-null (i.e. valid).
The same convention is now used for array data returned as a result
of read/scan operation.

The Python client has been updated correspondingly.

This patch also contains a few new test scenarios to cover the updated
functionality.

Change-Id: I6f0284a1cdf23234d30d18ae4b9f6ee3a9bd82f4
Reviewed-on: http://gerrit.cloudera.org:8080/23568
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>