KUDU-1261 use bits not bytes for validity in array1d.fbs
This changelist switches to using one-validity-bit-per-array-element
approach in encoding one-dimensional arrays.  It takes care both of
the C++ and the Java side, so now they both speak the same language
w.r.t. the array element validity.

I added a few new sub-scenarios to cover some of the edge cases with
the new approach, but the existing tests also provide good enough
coverage for this update.

Change-Id: I09d281f7e76f6ead3fac8cb853eea503b8fb0bfe
Reviewed-on: http://gerrit.cloudera.org:8080/23571
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>