KUDU-1261 fix criteria to select proper block writer
This fixed a bug in the criteria to select a proper method for writing
a particular column in MultiColumnWriter::AppendBlock().  The issue
would manifest itself when flushing memrowset containing a non-nullable
array column.  I'm not including tests in this particular changelist
because there is one already in a pending changelist [1] and in a
follow-up changelist including end-to-end tests for Kudu C++ client [2].

This is a follow-up to 9a1c19fd0853aef54bdd52af69194ee43f2a33b6.

[1] https://gerrit.cloudera.org/#/c/23565/
[2] https://gerrit.cloudera.org/#/c/23220/

Change-Id: I7041cd6a8c6f1bbcefacdb7cd3821e96774a60a2
Reviewed-on: http://gerrit.cloudera.org:8080/23589
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>