KUDU-1261 fix reading non-nullable array columns
This fixes a bug where there is an attempt to set non-null bit
into ColumnBlock of a non-nullable array column.  Since the bitmap
in such a ColumnBlock is nullptr, kudu-tserver process would crash
while trying to do so.

I'm not including tests in this particular changelist because there is
one already in a pending changelist [1] and I'm about to add a few more
into a follow-up changelist with end-to-end tests for Kudu C++ client [2].

This is a follow-up to 9a1c19fd0853aef54bdd52af69194ee43f2a33b6.

[1] https://gerrit.cloudera.org/#/c/23565/
[2] https://gerrit.cloudera.org/#/c/23220/

Change-Id: I85d2fa0b4f72546cb6e6213b5cc8eac526f0de3c
Reviewed-on: http://gerrit.cloudera.org:8080/23591
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>