[test] Skip test if hostname has no DNS record
With IPv6 support, some tests, that make use of DNS resolution,
can end up with error if the ip_config_mode is set to a family for
which 'hostname' has no corresponding DNS record.
For example, if a hostname resolves to just an IPv4 address on a system,
running the test with ip_config_mode as 'IPv6' fails because
getaddrinfo doesn't find any DNS record for the provided hostname with
desired family set to 'IPv6'.

This patch avoids such scenarios by, proactively, checking whether DNS
record exists or not and skips the test if it doesn't.

Change-Id: Id377c73a752e18324dbd83aad7e613b3ba99ee6b
Reviewed-on: http://gerrit.cloudera.org:8080/23611
Reviewed-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Tested-by: Marton Greber <greber.mx@gmail.com>