[java] mention the -DshowTestOutput flag in README
The `showTestOutput` option [1] is a shortcut to change the default
behavior of storing the logs when running the Gradle's 'test' task.

This changelist mentions it in the README.adoc file, so now it's easier
to find when necessary.

[1] https://github.com/apache/kudu/commit/7cdc9f7c7

Change-Id: I40d95279cb28729015aef837e99d0693a9114741
Reviewed-on: http://gerrit.cloudera.org:8080/23617
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
(cherry picked from commit b0c0059aecea5a977a7490a65b7d0840ed5bed76)
Reviewed-on: http://gerrit.cloudera.org:8080/23663