[docs] refresh `kudu master add` documentation
This changelist updates the documentation on using the `kudu master add`
CLI tool (which is effectively obsolete starting with Kudu 1.16.0).
It turned out the original version was quite misleading.  There is an
evidence that it confused users who attempted it because of the
omission of a few crucial points, especially if running the procedure
in a secure (i.e. Kerberized) Kudu cluster.

This changelist addresses the issue, at least clarifying on the
essential points of the procedure.  At the time of writing, the current
Kudu version is 1.18.0, and the `kudu master add` CLI tool is obsolete.
However, recently I found that there are still Kudu clusters of the
1.15.0 version in the wild, so I thought it makes sense to update the
docs on the procedure to avoid confusing and misleading users who might
try applying it at their 1.15.0 Kudu clusters.

This is a follow-up to 336c65bfd84a5115dc5c6be6521751d53aa0c286.

Change-Id: I8d2ae84d596a0c3ffe840cc16d3ac0d0d28d05bc
Reviewed-on: http://gerrit.cloudera.org:8080/23649
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
(cherry picked from commit 26776e78010239732879677123295b02dac7d92a)
Reviewed-on: http://gerrit.cloudera.org:8080/23659