[util] remove chromium-based Atomics from env_posix.cc
With this patch, Atomic64 and Barrier_AtomicIncrement() are no longer
used in env_posix.cc: the logic of PosixEnv::gettid() is now implemented
using std::atomic.  Also, the global static and thread-local static
variables are moved inside of the PosixEnv::gettid() method which is an
improvement as well, according to Clang-Tidy.

In addition, I added a new test scenario to verify that the generated
thread identifiers are indeed unique.  I didn't want to rely on existing
CHECK()/DCHECK() assertions in various places for indirect verification
of a few explicit invariants.

Apart from the new test scenario, there are no functional modifications
in this changelist.

Change-Id: I63d883de66c51d5d4a6c7eb1eebe63713f1f6dcc
Reviewed-on: http://gerrit.cloudera.org:8080/23689
Tested-by: Kudu Jenkins
Reviewed-by: Ashwani Raina <araina.git@gmail.com>
Reviewed-by: Marton Greber <greber.mx@gmail.com>