KUDU-1457 [7/n] Use correct URL host for ip mode
This patch uses correct URL host for a given IP config mode. If
webserver is running on wildcard address, tests using URL as input
to FetchURL need to make use of a corresponding loopback address.
For example, for a webserver running on '0.0.0.0', URL host is
'127.0.0.1' and similarly, for a webserver running on '[::]', URL
host is '[::1]'.
 * Based on IP config mode, choose correct service principal name.
 * Ensure test certificate contains subject name that matches with
   target hostname for non-IPv4 cases (i.e. when IP address is ::1).
   For this, new certificate and corresponding key is added with same
   configuration with additional subject name to accomodate '::1'.
 * Add a few more tests to IP config coverage.

Change-Id: I152b0755f286edce1738683b5d2cdf5804f4ff2a
Reviewed-on: http://gerrit.cloudera.org:8080/23711
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Tested-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Alexey Serbin <alexey@apache.org>