KUDU-2501 more instrumentation for TestNonRandomWorkloadLoadgen
This patch adds a bit of instrumentation to help with troubleshooting
of the ToolTest.TestNonRandomWorkloadLoadgen scenario when it fails.

Also, with this patch now all the test table's tablets are verified
to have zero bloom filter lookups after the test workload completes:
this is done by using '*' pattern for tablet identifier instead of
'nullptr' in the original version.  The original version verified only
the very first tablet in the metrics output, and that wasn't correct.

Change-Id: Ia9b4a63a00ecbf9af9151fef8eead10c627688c5
Reviewed-on: http://gerrit.cloudera.org:8080/23708
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>