KUDU-3723 fix interpretation of new range partition specs
In the context of KUDU-2671, a new notation for specifying hash schema
of a range partition has been added.  With the introduction of
range-specific hash schemas, the client side is free to send in
information on the newly added range partition as if it had a custom
hash schema, even if the new range's hash schema is the same as the
table-wide one.  However, the system catalog should have normalized
the provided information, and stored the information on ranges with
the table-wide and custom hash schemas differently.  The normalization
part was missing, and that was the reason for the reported issue.

This changelist addresses the problem and adds a new test scenario that
triggers an assertion without the fix.

Change-Id: Icceb138a919cd7afb572c6dd74695a3fcaaac99e
Reviewed-on: http://gerrit.cloudera.org:8080/23778
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>