[codegen] encode projection key once
There is no need to encode row projection key multiple times: one time
when scheduling a compilation task, another time when running it, and
one more time when inserting a new entry into the codegen cache.

The latter two steps were done by the same thread in the codegen
compilation pool, and there is just a single thread in the pool.
Apparently, it's better to encode the key only once by the thread which
schedules the compilation task upon requesting the corresponding row
projection.

With this update, there isn't a need for JITWrapper::EncodeOwnKey()
method anymore.

Change-Id: Ief2a1cf184c86c33031f90d92a10755da3e86e0b
Reviewed-on: http://gerrit.cloudera.org:8080/23819
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Zoltan Martonka <zmartonka@cloudera.com>
Reviewed-by: Ashwani Raina <araina.git@gmail.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>